[Arguments] Merge bleeding_edge revisions (8170, 8229] into arguments branch.
TBR=kmillikin@chromium.org
Review URL: http://codereview.chromium.org/7134068

git-svn-id: http://v8.googlecode.com/svn/branches/experimental@8269 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ arguments/samples/process.cc
@@ -535,7 +535,7 @@ void ParseOptions(int argc,
                   string* file) {
   for (int i = 1; i < argc; i++) {
     string arg = argv[i];
-    int index = arg.find('=', 0);
+    size_t index = arg.find('=', 0);
     if (index == string::npos) {
       *file = arg;
     } else {
